Switch background colors
Created the function selectColor and put it in the functions forward and back.

Set the session storage of color to a random number in between 1 and 6.

This function increases color by a random number in between 1 and 5 to ensure that it never picks the same random number twice in a row.

If color is greater than 6, it generates a new random number in between 1 and 6.