Text colors
Set the text color to light black.

Wrapped #quote and #author in the div #text.

Gave #text a white border.

Gave the body a linear gradient background.

Gave #quotesCont a darker linear gradient background of the same color.

Gave #text a lighter solid background of the same color.

Wrapped #text in the div #textCont.

Gave #textCont a white background so the background color of #quotesCont did not show through it.